- [*] refactor varcaser to egVar
