+++ easygen.go
@@ -137,15 +137,6 @@ func ParseFiles(HTML bool, filenames ...string) (Template, error) {
 		"year4":    year4,
 		"replace":  replace,
 		"replacec": replacec,
-		"cls2lc":   cls2lc.String,
-		"cls2uc":   cls2uc.String,
-		"cls2ss":   cls2ss.String,
-		"ck2lc":    ck2lc.String,
-		"ck2uc":    ck2uc.String,
-		"ck2ls":    ck2ls.String,
-		"ck2ss":    ck2ss.String,
-		"clc2ss":   clc2ss.String,
-		"cuc2ss":   cuc2ss.String,
 	})
 
 	if len(Opts.TemplateStr) > 0 {
