+++ template.go
@@ -38,15 +38,6 @@ var egFuncMap = FuncMap{
 	"minus1": minus1,
 	"dateI":  dateI,
 	"year4":  year4,
-	"cls2lc": cls2lc.String,
-	"cls2uc": cls2uc.String,
-	"cls2ss": cls2ss.String,
-	"ck2lc":  ck2lc.String,
-	"ck2uc":  ck2uc.String,
-	"ck2ls":  ck2ls.String,
-	"ck2ss":  ck2ss.String,
-	"clc2ss": clc2ss.String,
-	"cuc2ss": cuc2ss.String,
 }
 
 ////////////////////////////////////////////////////////////////////////////
